pyproject.toml
[tool.poetry]
name = "honeybot"
version = "0.1.0"
description = ""
authors = ["evgeniy <Evgexac@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.24"
loguru = "^0.6.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
